Restore test files
